Revert "Add in memory keyring."
